OY-3894 : Yhdistetty modalmachine
