+++ src/main/app/src/pages/HomePage/CopyConfirmationModal.tsx
@@ -3,92 +3,13 @@ import React, { useCallback, useEffect, useMemo } from 'react';
 import { useActor, useInterpret, useSelector } from '@xstate/react';
 import _fp from 'lodash/fp';
 import { useTranslation } from 'react-i18next';
-import { createMachine, spawn, actions } from 'xstate';
 
 import Modal from '#/src/components/Modal';
 import { Box, Button } from '#/src/components/virkailija';
 import { useContextOrThrow } from '#/src/hooks/useContextOrThrow';
+import { ModalMachine } from '#/src/machines/modalMachine';
 
 import { EntityListTable } from './EntitySearchList';
-import { EntitySelectionMachine } from './entitySelectionMachine';
-
-const { pure, assign, send } = actions;
-
-interface OpenEvent {
-  type: 'OPEN';
-}
-
-interface CloseEvent {
-  type: 'CLOSE';
-}
-
-type EntitySelection = Record<string, Record<string, any>>;
-
-interface SetEntitiesEvent {
-  type: 'SET_ENTITIES';
-  entities: EntitySelection;
-}
-
-interface CopyConfirmationMachineContext {
-  selectionRef: any;
-  entities?: EntitySelection;
-}
-
-export const CopyConfirmationModalMachine = createMachine(
-  {
-    schema: {
-      events: {} as OpenEvent | CloseEvent | SetEntitiesEvent,
-      context: {
-        entities: {} as EntitySelection,
-        selectionRef: null as any,
-      },
-    },
-    context: {
-      entities: {}, // Kaikki valittavissa olevat entiteetit
-      selectionRef: null,
-    },
-    initial: 'initializing',
-    states: {
-      initializing: {
-        entry: assign({
-          selectionRef: () => spawn(EntitySelectionMachine),
-        }),
-        always: {
-          target: 'closed',
-        },
-      },
-      open: {
-        entry: 'selectAll',
-        on: {
-          CLOSE: 'closed',
-        },
-      },
-      closed: {
-        on: {
-          OPEN: 'open',
-        },
-      },
-    },
-    on: {
-      SET_ENTITIES: {
-        actions: 'setEntities',
-      },
-    },
-  },
-  {
-    actions: {
-      setEntities: assign<CopyConfirmationMachineContext, any>({
-        entities: (ctx, e) => e.entities,
-      }),
-      selectAll: pure<CopyConfirmationMachineContext, any>(ctx => {
-        return send(
-          { type: 'RESET_SELECTION', items: ctx.entities },
-          { to: ctx.selectionRef }
-        );
-      }),
-    } as any,
-  }
-);
 
 export const CopyConfirmationModalContext = React.createContext({} as any);
 
@@ -114,7 +35,7 @@ export const useModalSelection = () => {
 };
 
 export const CopyConfirmationWrapper = ({ entities, children }) => {
-  const modalService = useInterpret(CopyConfirmationModalMachine);
+  const modalService = useInterpret(ModalMachine);
 
   useEffect(() => {
     modalService.send({ type: 'SET_ENTITIES', entities });
